chore: move 404 to standard layout
